test-scenarios/05.1-new-flaky-test.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Flaky test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(function() {
        assert_true(Math.random() > .5);
      }, "Intentionally flaky test");
   </script>
  </body>
</html>
